Moved Process unit tests to separate file and added additional tests
