+++ tests/unit/test_model.py
@@ -2,7 +2,6 @@
 # Copyright 2022 Atlan Pte. Ltd.
 import json
 from datetime import datetime
-from hashlib import md5
 from inspect import signature
 from pathlib import Path
 from unittest.mock import create_autospec
@@ -12,7 +11,6 @@ import pytest
 # from deepdiff import DeepDiff
 from pydantic.error_wrappers import ValidationError
 
-import pyatlan.cache.classification_cache
 from pyatlan.model.assets import (
     SQL,
     AccessControl,
@@ -34,9 +32,7 @@ from pyatlan.model.assets import (
     AuthPolicyValiditySchedule,
     AwsTag,
     AzureTag,
-    Badge,
     BadgeCondition,
-    Catalog,
     Column,
     ColumnProcess,
     ColumnValueFrequencyMap,
@@ -1418,262 +1414,3 @@ def test_create_readme(asset, content, asset_name, expected_name):
     assert readme.name == f"{expected_name} Readme"
     assert readme.attributes.asset == asset
     assert readme.description == content
-
-
-@pytest.mark.parametrize(
-    "name, connection_qualified_name, process_id, inputs,outputs, parent, message",
-    [
-        (None, "133/s3", None, [Catalog()], [Catalog()], None, "name is required"),
-        (
-            "bob",
-            None,
-            None,
-            [Catalog()],
-            [Catalog()],
-            None,
-            "connection_qualified_name is required",
-        ),
-        ("bob", "133/s3", None, None, [Catalog()], None, "inputs is required"),
-        (
-            "bob",
-            "133/s3",
-            None,
-            [],
-            [Catalog()],
-            None,
-            "inputs cannot be an empty list",
-        ),
-        ("bob", "133/s3", None, [Catalog()], None, None, "outputs is required"),
-        (
-            "bob",
-            "133/s3",
-            None,
-            [Catalog()],
-            [],
-            None,
-            "outputs cannot be an empty list",
-        ),
-    ],
-)
-def test_process_attributes_generate_qualified_name_without_required_parameter_raises_value_error(
-    name, connection_qualified_name, process_id, inputs, outputs, parent, message
-):
-    with pytest.raises(ValueError, match=message):
-        Process.Attributes.generate_qualified_name(
-            name=name,
-            connection_qualified_name=connection_qualified_name,
-            process_id=process_id,
-            inputs=inputs,
-            outputs=outputs,
-            parent=parent,
-        )
-
-
-@pytest.mark.parametrize(
-    "name, connection_qualified_name, process_id, inputs,outputs, parent, expected_value",
-    [
-        (
-            "doit",
-            "default/s3/1678379436102",
-            "123",
-            [Catalog()],
-            [Catalog()],
-            None,
-            "default/s3/1678379436102/123",
-        ),
-        (
-            "doit",
-            "default/s3/1678379436102",
-            None,
-            [Catalog(guid="123")],
-            [Catalog(guid="456")],
-            None,
-            "doitdefault/s3/1678379436102123456",
-        ),
-        (
-            "doit",
-            "default/s3/1678379436102",
-            None,
-            [Catalog(guid="456")],
-            [Catalog(guid="789")],
-            Catalog(guid="123"),
-            "doitdefault/s3/1678379436102123456789",
-        ),
-    ],
-)
-def test_process_attributes_generate_qualified_name(
-    name, connection_qualified_name, process_id, inputs, outputs, parent, expected_value
-):
-    if not process_id:
-        expected_value = md5(expected_value.encode()).hexdigest()
-
-    assert (
-        Process.Attributes.generate_qualified_name(
-            name=name,
-            connection_qualified_name=connection_qualified_name,
-            process_id=process_id,
-            inputs=inputs,
-            outputs=outputs,
-            parent=parent,
-        )
-        == expected_value
-    )
-
-
-@pytest.mark.parametrize(
-    "name, connection_qualified_name, process_id, inputs,outputs, parent, message",
-    [
-        (None, "133/s3", None, [Catalog()], [Catalog()], None, "name is required"),
-        (
-            "bob",
-            None,
-            None,
-            [Catalog()],
-            [Catalog()],
-            None,
-            "connection_qualified_name is required",
-        ),
-        ("bob", "133/s3", None, None, [Catalog()], None, "inputs is required"),
-        (
-            "bob",
-            "133/s3",
-            None,
-            [],
-            [Catalog()],
-            None,
-            "inputs cannot be an empty list",
-        ),
-        ("bob", "133/s3", None, [Catalog()], None, None, "outputs is required"),
-        (
-            "bob",
-            "133/s3",
-            None,
-            [Catalog()],
-            [],
-            None,
-            "outputs cannot be an empty list",
-        ),
-    ],
-)
-def test_process_attributes_create_without_required_parameter_raises_value_error(
-    name, connection_qualified_name, process_id, inputs, outputs, parent, message
-):
-    with pytest.raises(ValueError, match=message):
-        Process.Attributes.create(
-            name=name,
-            connection_qualified_name=connection_qualified_name,
-            process_id=process_id,
-            inputs=inputs,
-            outputs=outputs,
-            parent=parent,
-        )
-
-
-@pytest.mark.parametrize(
-    "name, connection_qualified_name, process_id, inputs,outputs, parent, expected_value",
-    [
-        (
-            "doit",
-            "default/s3/1678379436102",
-            "123",
-            [Catalog()],
-            [Catalog()],
-            None,
-            "default/s3/1678379436102/123",
-        ),
-        (
-            "doit",
-            "default/s3/1678379436102",
-            None,
-            [Catalog(guid="123")],
-            [Catalog(guid="456")],
-            None,
-            "doitdefault/s3/1678379436102123456",
-        ),
-        (
-            "doit",
-            "default/s3/1678379436102",
-            None,
-            [Catalog(guid="456")],
-            [Catalog(guid="789")],
-            Catalog(guid="123"),
-            "doitdefault/s3/1678379436102123456789",
-        ),
-    ],
-)
-def test_process_attributes_create(
-    name, connection_qualified_name, process_id, inputs, outputs, parent, expected_value
-):
-    if not process_id:
-        expected_value = md5(expected_value.encode()).hexdigest()
-
-    process = Process.create(
-        name=name,
-        connection_qualified_name=connection_qualified_name,
-        process_id=process_id,
-        inputs=inputs,
-        outputs=outputs,
-        parent=parent,
-    )
-
-    assert process.name == name
-    assert process.connection_qualified_name == connection_qualified_name
-    assert process.qualified_name == expected_value
-    assert process_id == process_id
-    assert process.inputs == inputs
-    assert process.outputs == outputs
-
-
-class Test_Badge_Attributes:
-    @pytest.mark.parametrize(
-        "name, cm_name, cm_attribute, badge_conditions, message",
-        [
-            (None, "Bob", "Dave", [BADGE_CONDITION], "name is required"),
-            ("Bob", None, "Dave", [BADGE_CONDITION], "cm_name is required"),
-            ("Bob", "", "Dave", [BADGE_CONDITION], "cm_name cannot be blank"),
-            ("Bob", "Dave", None, [BADGE_CONDITION], "cm_attribute is required"),
-            ("Bob", "Dave", "", [BADGE_CONDITION], "cm_attribute cannot be blank"),
-            ("Bob", "tom", "Dave", None, "badge_conditions is required"),
-            ("Bob", "tom", "Dave", [], "badge_conditions cannot be an empty list"),
-        ],
-    )
-    def test_create_when_required_parameters_are_missing_raises_value_error(
-        self, name, cm_name, cm_attribute, badge_conditions, message
-    ):
-        with pytest.raises(ValueError, match=message):
-            Badge.Attributes.create(
-                name=name,
-                cm_name=cm_name,
-                cm_attribute=cm_attribute,
-                badge_conditions=badge_conditions,
-            )
-
-    def test_create(self, monkeypatch):
-        def get_attr_id_for_name(set_name: str, attr_name: str):
-            return CM_ATTR_ID
-
-        def get_id_for_name(value):
-            return CM_ID
-
-        monkeypatch.setattr(
-            pyatlan.cache.custom_metadata_cache.CustomMetadataCache,
-            "get_id_for_name",
-            get_id_for_name,
-        )
-
-        monkeypatch.setattr(
-            pyatlan.cache.custom_metadata_cache.CustomMetadataCache,
-            "get_attr_id_for_name",
-            get_attr_id_for_name,
-        )
-
-        badge = Badge.Attributes.create(
-            name="bob",
-            cm_name="Monte Carlo",
-            cm_attribute="dummy",
-            badge_conditions=[BADGE_CONDITION],
-        )
-        assert badge.name == "bob"
-        assert badge.qualified_name == f"badges/global/{CM_ID}.{CM_ATTR_ID}"
-        assert badge.badge_metadata_attribute == f"{CM_ID}.{CM_ATTR_ID}"
-        assert badge.badge_conditions == [BADGE_CONDITION]
